ASSIGNMENT_02/hello-agent/pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MUDASSIR ABRAR", email = "161126723+MudassarAbrar@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
hello-agent = "hello_agent:main"
my-command = "hello_agent.agent_hello:my_first_agent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
